test(Auth-API): allow stub of auth middleware
