+++ test/testConfig.js
@@ -1,13 +1,10 @@
-const AuthMiddleware = require('../src/middleware/auth.middleware');
 const sinon = require('sinon');
 
-const realAuthMiddleware = AuthMiddleware.auth;
-const stubAuthMiddleware = sinon.stub(AuthMiddleware, 'auth').callsFake((roles) => (req, res, next) => next());
-
-const app = require("../src/server");
+const AuthMiddleware = require('../src/middleware/auth.middleware');
+const stubbedAuthMiddleware = sinon.stub(AuthMiddleware, 'auth').callsFake((roles) => {
+    return (req, res, next) => next();
+});
 
 module.exports = {
-    realAuthMiddleware,
-    stubAuthMiddleware,
-    app
+    stubbedAuthMiddleware
 };
